Move Tvh5 functions to Tvh5Format class
Instead of writing and reading the Tvh5 format inside the EmdFormat,
move the Tvh5 format functions into its own class: Tvh5Format.

While migrating the functions over, EmdFormat was also cleaned up a
little bit.
